Added ability for site Admin to reclac standings on demand
Normally it happens every time the master bracket is saved. Makes sense once the tourney starts but max scores wouldn't be calculated before games start and people will probably want to see that. Probably a better way to do it but for now we could just run this every so often between now and Thursday.